fix: se agrego material icons
